Full Back

+++ src/main/java/co/edu/usbcali/aerolinea/repository/AsientoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AsientoRepository extends JpaRepository<Asiento, Long> {
+public interface AsientoRepository extends JpaRepository<Asiento, Integer> {
 }
