+++ src/main/java/co/edu/usbcali/aerolinea/repository/AvionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvionRepository extends JpaRepository<Avion, Long> {
+public interface AvionRepository extends JpaRepository<Avion, Integer> {
 
 }
