+++ src/main/java/co/edu/usbcali/aerolinea/repository/TipoAsientoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoAsientoRepository extends JpaRepository<TipoAsiento, Long> {
+public interface TipoAsientoRepository extends JpaRepository<TipoAsiento, Integer> {
 }
